finishes upgrade to new form action and improves draggable in form editing
The form action is upgraded to the /results/[:id] when a form is created or copied, using a helper method in the form model.  Draggable was causing field elements to move wildly out of place when a field was deleted, so when a field is deleted a ghost box of the same dimensions is left in its place to keep fields from being rearranged automatically.