Phase 2a of new form edit process complete
An edit box containing the editing view for the selected field
appears below the field that is being edited.  This edit box
is appended to the form and positioned below the field in focus.
The app needs some adjustments to make the edit box function
like it did before when it was in the sidebar, and needs some
CSS updating. Additional buttons will be enabled, allowing for
saving the form, duplicating the field, deleting the field,
and unfocusing the field.