Phase 2b of new form edit process complete
The window is scrolled to center the edit box and edited
field. The edit box can sync with the field, making and
receiving changes, but not perfectly.  3 out of 4 of the
edit box buttons are working (Save Form, Delete Field,
and Done editing).  Still need to be able to make a copy
of the field and go through the field types and update
each to ensure they sync properly with the edit box.