Fixes Chrome glitch for forms index previews, starts results view
Chrome was making the form title disappear on the form preview
once it had been clicked in the forms index, but only for
some forms, apparently depending on the form content. It
appears to be a Chrome specific bug, the form preview was
simplified in structure to avoid it.  The move handle in
the form edit view is bigger for better UX on touchscreens.