add emails and notify_by columns to form table
notify_by is an object containing strings of which types
of notifications should be made when the form is submitted.
emails is an object containing the emails that should be
used if the results should be emailed. The Backbone Form
model was extended with initialize, parse, and toJSON methods
to make sure these objects are sent to the server as strings
and then stored in the backbone model as objects. Form Settings
view was modified to use emails and notify_by as objects.